- Añadida la clase MySqlDataset. - Añadido el script de inicialización de la base de datos en MySQL. - Eliminado el atributo connection_ de la clase PostgreSqlDataset. Ahora, en cada operación se crea una conexión nueva. - Modificado el tipo de datos del atributo size_ de la clase Dataset. Ahora es un long unsigned int. - Actualizada la documentación del código fuente. - Modificado el Makefile y el programa de prueba para compilar y depurar la clase MySqlDataset.

+++ trunk/inc/Dataset.hpp
@@ -51,7 +51,7 @@ class Dataset
 		///	@brief	Get the number of samples in the dataset.
 		///
 		///	@return	Number of samples in the dataset.
-		virtual const unsigned int& size () const;
+		virtual const long unsigned int& size () const;
 
 		///	@brief	Get the number of features per sample.
 		///
@@ -92,7 +92,7 @@ class Dataset
 
 		std::map<std::string, unsigned int>	classes_;	///< Map of classes that associates a character with its code.
 
-		unsigned int						size_;		///< Number of samples.
+		long unsigned int					size_;		///< Number of samples.
 
 		unsigned int						features_;	///< Number of features per sample.
 };
@@ -103,7 +103,7 @@ inline const Sample& Dataset::at (const unsigned int& n) const
 	return samples_.at(n);
 };
 
-inline const unsigned int& Dataset::size () const
+inline const long unsigned int& Dataset::size () const
 {
 	return size_;
 };
